feat: move DiscoverCommunityFeedFilter to commons
